Fix line-unbreaking and decoding of log files
In the line-unbreaking heuristics, use the fact that tex breaks lines of
more than 79 bytes, not chars -- with unicode, these are different.
Also, the log file is encoded differently from the console encoding
reported by oemcp. For now try UTF-8 only; in the future, sniff the
encoding using the usual heuristics.